Rename build runtime to nodejs
